WooCommerce Install: Update styles
Brings styles closer to the figma
