+++ client/signup/steps/woocommerce-install/confirm/index.tsx
@@ -1,5 +1,4 @@
 import { isEnabled } from '@automattic/calypso-config';
-import { NextButton } from '@automattic/onboarding';
 import styled from '@emotion/styled';
 import { useI18n } from '@wordpress/react-i18n';
 import page from 'page';
@@ -13,21 +12,11 @@ import WarningCard from 'calypso/components/warning-card';
 import StepWrapper from 'calypso/signup/step-wrapper';
 import { submitSignupStep } from 'calypso/state/signup/progress/actions';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import { ActionSection, StyledNextButton } from '..';
 import SupportCard from '../components/support-card';
 import useWooCommerceOnPlansEligibility from '../hooks/use-woop-handling';
 import type { WooCommerceInstallProps } from '../';
 import './style.scss';
-
-export const ActionSection = styled.div`
-	display: flex;
-	justify-content: space-between;
-	align-items: baseline;
-	flex-wrap: wrap;
-
-	@media ( max-width: 320px ) {
-		align-items: center;
-	}
-`;
 
 const Divider = styled.hr`
 	border-top: 1px solid #eee;
@@ -37,13 +26,6 @@ const Divider = styled.hr`
 
 const WarningsOrHoldsSection = styled.div`
 	margin-bottom: 40px;
-`;
-
-export const StyledNextButton = styled( NextButton )`
-	@media ( max-width: 320px ) {
-		width: 100%;
-		margin-bottom: 20px;
-	}
 `;
 
 export default function Confirm( props: WooCommerceInstallProps ): ReactElement | null {
