+++ client/signup/steps/woocommerce-install/confirm/style.scss
@@ -1,106 +1,1 @@
-@import '@automattic/onboarding/styles/mixins';
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
-
-body.is-section-signup .is-woocommerce-install {
-	.signup__step.is-business-info, // @todo: tidy all steps styles
-	.signup__step.is-store-address, // @todo: tidy all steps styles
-	.signup__step.is-confirm {
-		min-height: 95vh;
-		display: flex;
-		align-items: center;
-
-		@media ( max-width: 660px ) {
-			min-height: 0;
-		}
-		// Two columns layout
-		.step-wrapper {
-			display: flex;
-
-			.step-wrapper__header {
-				@media ( min-width: 660px ) {
-					margin-top: 0 !important;
-				}
-
-				.formatted-header {
-					text-align: inherit;
-					@media ( min-width: 1024px ) {
-						margin-left: 24px;
-					}
-
-					.formatted-header__title {
-						text-align: inherit;
-						@media ( max-width: 660px ) {
-							padding: 0 10px 0 0;
-
-						}
-
-					}
-
-					.formatted-header__subtitle {
-						@media ( max-width: 660px ) {
-							margin-top: 8px;
-							padding: 0 10px 0 0;
-						}
-
-						margin-top: 20px;
-						text-align: inherit;
-						max-width: 448px;
-					}
-				}
-			}
-		}
-
-		.is-wide-layout {
-			max-width: 1024px;
-			display: grid;
-			grid-template-columns: 1fr 1fr;
-			align-items: flex-start;
-			justify-content: center;
-			column-gap: 2%;
-
-			@media ( max-width: 660px ) {
-				max-width: 90vw;
-				grid-template-columns: 1fr;
-			}
-		}
-
-		.step-wrapper__content {
-			display: flex;
-			margin: 0 24px;
-
-			.step-business-info__instructions-container, // @todo: tidy all steps styles
-			.step-store-address__instructions-container, // @todo: tidy all steps styles
-			.confirm__instructions-container {
-				font-size: 0.875rem;
-				letter-spacing: -0.16px;
-				color: var( --studio-gray-60 );
-				line-height: 2rem;
-				flex-direction: column;
-				max-width: 520px;
-				min-width: 380px;
-
-				@media ( max-width: 660px ) {
-					max-width: 90vw;
-				}
-
-				p {
-					margin-top: 16px;
-				}
-
-				.components-text-control__input {
-					box-sizing: border-box;
-				}
-			}
-		}
-	}
-}
-
-.confirm__info-section {
-	flex: 1;
-
-	.wpcom__loading-ellipsis {
-		display: block;
-		margin: 0 auto;
-	}
-}
+@import '../style';
