+++ client/signup/steps/woocommerce-install/index.tsx
@@ -1,3 +1,5 @@
+import { NextButton } from '@automattic/onboarding';
+import styled from '@emotion/styled';
 import type { GoToStep, GoToNextStep } from '../../types';
 
 export interface WooCommerceInstallProps {
@@ -17,3 +19,22 @@ export interface WooCommerceInstallProps {
 		siteConfirmed?: number;
 	};
 }
+
+export const ActionSection = styled.div`
+	display: flex;
+	justify-content: space-between;
+	align-items: baseline;
+	flex-wrap: wrap;
+	margin-top: 40px;
+
+	@media ( max-width: 320px ) {
+		align-items: center;
+	}
+`;
+
+export const StyledNextButton = styled( NextButton )`
+	@media ( max-width: 320px ) {
+		width: 100%;
+		margin-bottom: 20px;
+	}
+`;
