+++ client/signup/steps/woocommerce-install/step-business-info/index.tsx
@@ -7,8 +7,8 @@ import StepWrapper from 'calypso/signup/step-wrapper';
 import { submitSignupStep } from 'calypso/state/signup/progress/actions';
 import { getSiteDomain } from 'calypso/state/sites/selectors';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import { ActionSection, StyledNextButton } from '..';
 import SupportCard from '../components/support-card';
-import { ActionSection, StyledNextButton } from '../confirm';
 import { useSiteSettings, WOOCOMMERCE_ONBOARDING_PROFILE } from '../hooks/use-site-settings';
 import type { WooCommerceInstallProps } from '..';
 import './style.scss';
