+++ client/signup/steps/woocommerce-install/step-business-info/style.scss
@@ -1,4 +1,1 @@
-@import '@automattic/onboarding/styles/mixins';
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
-
+@import '../style';
