+++ client/signup/steps/woocommerce-install/step-store-address/index.tsx
@@ -11,8 +11,8 @@ import { fetchWooCommerceCountries } from 'calypso/state/countries/actions';
 import getCountries from 'calypso/state/selectors/get-countries';
 import { submitSignupStep } from 'calypso/state/signup/progress/actions';
 import { getSelectedSiteId } from 'calypso/state/ui/selectors';
+import { ActionSection, StyledNextButton } from '..';
 import SupportCard from '../components/support-card';
-import { ActionSection, StyledNextButton } from '../confirm';
 import {
 	useSiteSettings,
 	WOOCOMMERCE_STORE_ADDRESS_1,
