+++ client/signup/steps/woocommerce-install/step-store-address/style.scss
@@ -1,23 +1,22 @@
-@import '@automattic/onboarding/styles/mixins';
-@import '@wordpress/base-styles/breakpoints';
-@import '@wordpress/base-styles/mixins';
+@import '../style';
 
 .step-store-address__instructions-container {
 	.components-combobox-control__suggestions-container {
 		// Remove `width: 100%;` https://github.com/WordPress/gutenberg/blob/0135921bcb72ceec629fa8f9e755891afab747b9/packages/components/src/combobox-control/style.scss#L32
 		width: unset;
 
-		// Remove margin from component to match text controls
+		// Remove margin from component to match text controls.
 		margin-bottom: unset;
 	}
 
 	.components-form-token-field__suggestions-list {
+		margin: 4px -16px -12px;
 		width: min-content;
 	}
 
 	.components-combobox-control__reset.components-button {
-		// Height of 22 ensures combobox input matches height of text controls
-		height: 22px;
+		// Height of 18 ensures combobox input matches height of text controls.
+		height: 18px;
 	}
 
 	.form-input-validation.is-error {
@@ -41,7 +40,7 @@
 		}
 
 		.components-base-control__field {
-			// Remove margin-bottom on error so errors are positioned closer to the corresponding field
+			// Remove margin-bottom on error so errors are positioned closer to the corresponding field.
 			margin-bottom: unset;
 		}
 	}
